Refactor remediation trait to streamline product ID handling.
Moved `get_all_product_ids` implementation from `getter_implementations.rs` to `getter_traits.rs` for better modularity and reuse. Removed redundant imports and optimized closures for clarity and consistency in all related modules.
