+++ csaf-lib/src/csaf/getter_traits.rs
@@ -1,4 +1,5 @@
 use crate::csaf::csaf2_1::schema::CategoryOfTheRemediation;
+use crate::csaf::helpers::resolve_product_groups;
 use std::collections::BTreeSet;
 
 pub trait CsafTrait {
@@ -6,6 +7,7 @@ pub trait CsafTrait {
     type ProductTreeType: ProductTreeTrait;
 
     fn get_product_tree(&self) -> Option<Self::ProductTreeType>;
+
     fn get_vulnerabilities(&self) -> Vec<Self::VulnerabilityType>;
 }
 
@@ -17,9 +19,27 @@ pub trait VulnerabilityTrait {
 
 pub trait RemediationTrait {
     fn get_category(&self) -> CategoryOfTheRemediation;
+
     fn get_product_ids(&self) -> Option<Vec<&String>>;
+
     fn get_group_ids(&self) -> Option<Vec<&String>>;
-    fn get_all_product_ids(&self, doc: &impl CsafTrait) -> Option<BTreeSet<String>>;
+
+    fn get_all_product_ids(&self, doc: &impl CsafTrait) -> Option<BTreeSet<String>> {
+        if self.get_product_ids().is_none() && self.get_group_ids().is_none() {
+            None
+        } else {
+            let mut product_set: BTreeSet<String> = match self.get_product_ids() {
+                Some(product_ids) => product_ids.iter().map(|p| p.to_string()).collect(),
+                None => BTreeSet::new(),
+            };
+            if let Some(product_groups) = self.get_group_ids() {
+                if let Some(product_ids) = resolve_product_groups(doc, product_groups) {
+                    product_set.extend(product_ids.iter().map(|p| p.to_string()));
+                }
+            }
+            Some(product_set)
+        }
+    }
 }
 
 pub trait ProductTreeTrait {
@@ -30,5 +50,6 @@ pub trait ProductTreeTrait {
 
 pub trait ProductGroupTrait {
     fn get_group_id(&self) -> &String;
+
     fn get_product_ids(&self) -> Vec<&String>;
 }
